+++ library/data_structures/lazy_segment_tree.hpp
