+++ library/data_structures/uncommon/disjoint_sparse_table.hpp
