+++ library/data_structures/uncommon/distinct_query.hpp
