+++ library/data_structures/uncommon/implicit_seg_tree.hpp
