+++ tests/.config/.cppcheck_suppression_list
@@ -10,7 +10,7 @@ unusedStructMember:../library/flow/hungarian.hpp:9
 functionConst:../kactl/content/data-structures/SubMatrix.h:24
 useStlAlgorithm:../kactl/content/numerical/FastFourierTransform.h:53
 constVariable:../kactl/content/number-theory/MillerRabin.h:17
-passedByValue:../library/range_data_structures/uncommon/implicit_seg_tree.hpp
+passedByValue:../library/data_structures/uncommon/implicit_seg_tree.hpp
 constParameter:../hackpack-cpp/content/number-theory/ModInt.h:29
 functionConst:../hackpack-cpp/content/graphs/TreeLifting.h:22
 functionConst:../hackpack-cpp/content/graphs/TreeLifting.h:31
