+++ tests/.config/hackpack.tex
@@ -187,17 +187,17 @@
 
 \lstinputlisting[caption={\textbf{RANGE DATA STRUCTURES}}, frame=none]{NULL}
 
-\lstinputlisting[language=C++, caption={Binary Indexed Tree}]{../library/range_data_structures/binary_indexed_tree.hpp}
-\lstinputlisting[language=C++, caption={Recursive Lazy Segment Tree}]{../library/range_data_structures/lazy_segment_tree.hpp}
-\lstinputlisting[language=C++, caption={Sparse Table}]{../library/range_data_structures/sparse_table.hpp}
+\lstinputlisting[language=C++, caption={Binary Indexed Tree}]{../library/data_structures/binary_indexed_tree.hpp}
+\lstinputlisting[language=C++, caption={Recursive Lazy Segment Tree}]{../library/data_structures/lazy_segment_tree.hpp}
+\lstinputlisting[language=C++, caption={Sparse Table}]{../library/data_structures/sparse_table.hpp}
 
 % uncommon
-\lstinputlisting[language=C++, caption={Disjoint Sparse Table}]{../library/range_data_structures/uncommon/disjoint_sparse_table.hpp}
-\lstinputlisting[language=C++, caption={Number Distinct Elements}]{../library/range_data_structures/uncommon/distinct_query.hpp}
-\lstinputlisting[language=C++, caption={Implicit Lazy Segment Tree}]{../library/range_data_structures/uncommon/implicit_seg_tree.hpp}
-\lstinputlisting[language=C++, caption={Iterative Lazy Segment Tree}]{../library/range_data_structures/uncommon/iterative_lazy_segment_tree.hpp}
-\lstinputlisting[language=C++, caption={Kth Smallest}]{../library/range_data_structures/uncommon/kth_smallest.hpp}
-\lstinputlisting[language=C++, caption={Merge Sort Tree}]{../library/range_data_structures/uncommon/merge_sort_tree.hpp}
+\lstinputlisting[language=C++, caption={Disjoint Sparse Table}]{../library/data_structures/uncommon/disjoint_sparse_table.hpp}
+\lstinputlisting[language=C++, caption={Number Distinct Elements}]{../library/data_structures/uncommon/distinct_query.hpp}
+\lstinputlisting[language=C++, caption={Implicit Lazy Segment Tree}]{../library/data_structures/uncommon/implicit_seg_tree.hpp}
+\lstinputlisting[language=C++, caption={Iterative Lazy Segment Tree}]{../library/data_structures/uncommon/iterative_lazy_segment_tree.hpp}
+\lstinputlisting[language=C++, caption={Kth Smallest}]{../library/data_structures/uncommon/kth_smallest.hpp}
+\lstinputlisting[language=C++, caption={Merge Sort Tree}]{../library/data_structures/uncommon/merge_sort_tree.hpp}
 
 \lstinputlisting[caption={\textbf{STRINGS}}, frame=none]{NULL}
 \lstinputlisting[language=C++, caption={Binary Trie}]{../library/strings/binary_trie.hpp}
