+++ tests/online_judge_tests/data_structures/disjoint_rmq_lines.test.cpp
@@ -1,7 +1,7 @@
 #define PROBLEM "https://judge.yosupo.jp/problem/queue_operate_all_composite"
 #include "../template.hpp"
 
-#include "../../../library/range_data_structures/uncommon/disjoint_sparse_table.hpp"
+#include "../../../library/data_structures/uncommon/disjoint_sparse_table.hpp"
 
 int main() {
 	cin.tie(0)->sync_with_stdio(0);
