+++ tests/online_judge_tests/data_structures/fenwick_tree_lower_bound.test.cpp
@@ -1,7 +1,7 @@
 #define PROBLEM "https://judge.yosupo.jp/problem/predecessor_problem"
 #include "../template.hpp"
 
-#include "../../../library/range_data_structures/binary_indexed_tree.hpp"
+#include "../../../library/data_structures/binary_indexed_tree.hpp"
 
 int main() {
 	cin.tie(0)->sync_with_stdio(0);
