+++ tests/online_judge_tests/data_structures/kth_smallest.test.cpp
@@ -1,7 +1,7 @@
 #define PROBLEM "https://judge.yosupo.jp/problem/range_kth_smallest"
 #include "../template.hpp"
 
-#include "../../../library/range_data_structures/uncommon/kth_smallest.hpp"
+#include "../../../library/data_structures/uncommon/kth_smallest.hpp"
 
 int main() {
 	cin.tie(0)->sync_with_stdio(0);
