+++ tests/online_judge_tests/data_structures/merge_sort_tree.test.cpp
@@ -4,7 +4,7 @@
 #undef _GLIBCXX_DEBUG
 #include "../template.hpp"
 
-#include "../../../library/range_data_structures/uncommon/merge_sort_tree.hpp"
+#include "../../../library/data_structures/uncommon/merge_sort_tree.hpp"
 
 int main() {
 	cin.tie(0)->sync_with_stdio(0);
