+++ tests/online_judge_tests/data_structures/sparse_table.test.cpp
@@ -1,7 +1,7 @@
 #define PROBLEM "https://judge.yosupo.jp/problem/staticrmq"
 #include "../template.hpp"
 
-#include "../../../library/range_data_structures/sparse_table.hpp"
+#include "../../../library/data_structures/sparse_table.hpp"
 
 int main() {
 	cin.tie(0)->sync_with_stdio(0);
